default.html, search.js and search.html changed
Lunr.js version has been updated. Code was changed to reflect this update (now the to add the posts, it is done inside the idx file). All of the code inside of the search.js file was moved into search.html script bracket to be near the fetching of posts from site.posts. lunr.min.js was deleted as it was an old version. The new versions of lunr.js is referenced in default.html (head).